# contracts/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Kanban API
  version: 1.1.0
...
paths:
  /api/admin/stats:
    get:
      operationId: Admin_getStats
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatsResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Admin
  /api/admin/users:
    get:
      operationId: Admin_listUsers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 1
          explode: false
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
          explode: false
        - name: search
          in: query
          required: false
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - totalCount
                  - page
                  - pageSize
                  - totalPages
                  - hasNextPage
                  - hasPreviousPage
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUserListItem'
                  totalCount:
                    type: integer
                    format: int32
                    description: Total number of records matching the query
                  page:
                    type: integer
                    format: int32
                    description: Current page number (1-based)
                  pageSize:
                    type: integer
                    format: int32
                    description: Number of items per page
                  totalPages:
                    type: integer
                    format: int32
                    description: Total number of pages
                  hasNextPage:
                    type: boolean
                    description: Whether there is a next page
                  hasPreviousPage:
                    type: boolean
                    description: Whether there is a previous page
                description: Generic paginated response wrapper
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Admin
  /api/admin/users/{userId}/role:
    patch:
      operationId: Admin_updateUserRole
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListItem'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
  /api/admin/users/{userId}/status:
    patch:
      operationId: Admin_updateUserStatus
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListItem'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
  /api/auth/login:
    post:
      operationId: Auth_login
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
  /api/auth/logout:
    post:
      operationId: Auth_logout
      parameters: []
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
  /api/auth/me:
    get:
      operationId: Auth_me
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
    patch:
      operationId: Auth_updateProfile
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
  /api/auth/register:
    post:
      operationId: Auth_register
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
...
security:
  - {}
  - ApiKeyAuth: []
components:
  schemas:
    AdminStatsResponse:
      type: object
      required:
        - totalUsers
        - activeUsers
        - adminCount
        - totalBoards
        - recentUsers
      properties:
        totalUsers:
          type: integer
          format: int32
        activeUsers:
          type: integer
          format: int32
        adminCount:
          type: integer
          format: int32
        totalBoards:
          type: integer
          format: int32
        recentUsers:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserListItem'
    AdminUserListItem:
      type: object
      required:
        - id
        - email
        - displayName
        - role
        - isActive
        - createdAt
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      description: Standard error response returned on failures
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - displayName
      properties:
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
...
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        currentPassword:
          type: string
        newPassword:
          type: string
        avatarUrl:
          type: string
...
    UpdateUserRoleRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
    UpdateUserStatusRequest:
      type: object
      required:
        - isActive
      properties:
        isActive:
          type: boolean
    UserResponse:
      type: object
      required:
        - id
        - email
        - displayName
        - role
        - createdAt
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
        avatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: session
servers:
  - url: http://localhost:5010
    description: Development server
    variables: {}